Extract AutoTagOption
