+++ front/app/containers/Admin/projects/project/analysis/Tags/AutoTaggingModal/Step1.tsx
@@ -1,17 +1,13 @@
-import React, { ReactNode } from 'react';
+import React from 'react';
 
 import {
   Box,
   Title,
   Text,
   colors,
-  Spinner,
-  IconTooltip,
   Icon,
   Label,
   Radio,
-  stylingConsts,
-  Tooltip,
 } from '@citizenlab/cl2-component-library';
 import { isEmpty } from 'lodash-es';
 import { useParams } from 'react-router-dom';
@@ -21,31 +17,14 @@ import useAnalysis from 'api/analyses/useAnalysis';
 import { AutoTaggingMethod } from 'api/analysis_background_tasks/types';
 import { IInputsFilterParams } from 'api/analysis_inputs/types';
 import useInfiniteAnalysisInputs from 'api/analysis_inputs/useInfiniteAnalysisInputs';
-import { TagType } from 'api/analysis_tags/types';
 
 import useFeatureFlag from 'hooks/useFeatureFlag';
 
 import { useIntl } from 'utils/cl-intl';
 
 import messages from '../messages';
-import Tag from '../Tag';
 
-const AutoTagMethodContainer = styled.div<{ isDisabled: boolean }>`
-  background-color: ${colors.grey100};
-  border-radius: 3px;
-  padding: 16px;
-  ${({ isDisabled }) =>
-    isDisabled
-      ? `opacity: 0.5;
-         cursor: not-allowed;`
-      : `
-      cursor: pointer;
-      &:hover {
-        border: 1px black;
-        box-shadow: rgba(0, 0, 0, 0.02) 0px 1px 3px 0px,
-          rgba(27, 31, 35, 0.15) 0px 0px 0px 1px;
-      }`}
-`;
+import AutoTagOption from './AutoTagOption';
 
 const AutoTagTargetContainer = styled.div`
   flex: 1;
@@ -67,97 +46,6 @@ const AutoTagTargetContainer = styled.div`
     pointer-events: none;
   }
 `;
-
-const AutoTagOption = ({
-  children,
-  tagType,
-  title,
-  onSelect,
-  isDisabled,
-  isLoading,
-  tooltip,
-  isRecommended,
-}: {
-  children: ReactNode;
-  tagType: TagType;
-  title: string;
-  onSelect: () => void;
-  isDisabled: boolean;
-  isLoading: boolean;
-  tooltip?: string;
-  isRecommended?: boolean;
-}) => {
-  const { formatMessage } = useIntl();
-  return (
-    <Box w="30%">
-      <Tooltip
-        content={
-          <p>{formatMessage(messages.advancedAutotaggingUpsellMessage)}</p>
-        }
-        zIndex={9999999}
-        disabled={!isDisabled}
-      >
-        <AutoTagMethodContainer
-          onClick={isDisabled || isLoading ? undefined : () => onSelect()}
-          isDisabled={isDisabled || isLoading}
-          tabIndex={0}
-        >
-          <Box
-            display="flex"
-            justifyContent="space-between"
-            alignItems="center"
-            mb="8px"
-          >
-            <Box w="32px">
-              <Tag tagType={tagType} name="&nbsp;" />
-            </Box>
-            {isRecommended && (
-              <Box
-                bgColor={colors.success}
-                py="4px"
-                px="8px"
-                borderRadius={stylingConsts.borderRadius}
-                display="flex"
-                gap="4px"
-                alignItems="center"
-              >
-                <Icon
-                  name="stars"
-                  fill={colors.white}
-                  width="16px"
-                  height="16px"
-                />
-                <Text color="white" m="0px" fontSize="s">
-                  {formatMessage(messages.recommended)}
-                </Text>
-              </Box>
-            )}
-            {isDisabled && <Icon name="lock" />}
-          </Box>
-          <Box
-            display="flex"
-            justifyContent="flex-start"
-            alignItems="center"
-            gap="6px"
-          >
-            <Title variant="h6" m="0px">
-              {title}
-            </Title>
-            {isLoading && (
-              <Box mx="16px">
-                <Spinner size="24px" />
-              </Box>
-            )}
-            {tooltip && <IconTooltip content={tooltip} icon="info-outline" />}
-          </Box>
-          <Text mt="12px" mb="0px">
-            {children}
-          </Text>
-        </AutoTagMethodContainer>
-      </Tooltip>
-    </Box>
-  );
-};
 
 type Props = {
   onSelectMethod: (tagType: AutoTaggingMethod) => void;
